requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
fastapi==0.110.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
joblib==1.3.2
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pika==1.3.2
pluggy==1.4.0
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==3.12.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
threadpoolctl==3.4.0
typing_extensions==4.10.0
tzdata==2024.1
unicode==2.9
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.27.1
